requirements.txt
absl-py==2.0.0
accelerate==0.21.0
addict==2.4.0
aenum==3.1.15
affine==2.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
aliyun-python-sdk-alimt==3.2.0
aliyun-python-sdk-core==2.13.10
altair==5.1.2
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
basicsr==1.4.2
beautifulsoup4==4.12.2
bidict==0.22.1
bitsandbytes==0.41.1
black==24.2.0
bleach==6.1.0
blendmodes==2022
blinker==1.7.0
blis==0.7.11
boltons==23.1.1
boto3==1.29.5
botocore==1.32.5
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
clip==1.0
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.3
configobj==5.0.8
contexttimer==0.3.3
contourpy==1.2.0
controlnet-aux==0.0.3
cryptography==41.0.5
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
decord==0.6.0
deprecation==2.1.0
develop==0.2.1
diffusers==0.16.1
dill==0.3.7
distlib==0.3.7
distro==1.8.0
docutils==0.20.1
einops==0.4.1
exceptiongroup==1.1.3
executing==2.0.1
facexlib==0.3.0
fairscale==0.4.4
fastapi==0.94.0
ffmpy==0.3.1
filelock==3.13.1
filterpy==1.4.5
Flask==2.2.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flaskwebgui==0.3.5
flatbuffers==23.5.26
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.32
google-auth==2.23.4
google-auth-oauthlib==0.4.6
gradio==3.41.2
gradio_client==0.5.0
grpcio==1.59.2
h11==0.12.0
hf_transfer==0.1.6
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.19.0
humanfriendly==10.0
hydra-core==1.3.2
identify==2.5.31
idna==3.4
image-reward==1.5
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
invisible-watermark==0.2.0
iopath==0.1.9
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonmerge==1.8.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kaggle==1.5.16
kaleido==0.2.1
kiwisolver==1.4.5
kornia==0.6.7
lama-cleaner==1.2.5
langcodes==3.3.0
lark==1.1.2
lazy_loader==0.3
lightning-utilities==0.9.0
llvmlite==0.41.1
lmdb==1.4.1
loguru==0.7.2
lpips==0.1.4
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.7
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nh3==0.2.15
ninja==1.11.1.1
nodeenv==1.8.0
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
onnxruntime==1.16.1
open-clip-torch==2.20.0
openai==1.3.4
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.10
packaging==23.2
pandas==2.1.2
parso==0.8.3
pathos==0.3.1
pathspec==0.12.1
picologging==0.9.3
piexif==1.1.3
Pillow==9.5.0
platformdirs==3.11.0
plotly==5.18.0
pooch==1.8.0
portalocker==2.8.2
pox==0.3.3
ppft==1.7.6.7
pre-commit==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.13
pydeck==0.8.1b0
pydub==0.25.1
PyExecJS==1.5.1
pyfunctional==1.4.3
Pygments==2.16.1
PyMatting==1.1.10
pyparsing==3.1.1
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-hglib==2.6.2
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
python-socketio==5.10.0
pytorch-lightning==1.9.4
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
rasterio==1.3.10
readme_renderer==43.0
realesrgan==0.3.0
referencing==0.30.2
regex==2023.10.3
rembg==2.0.38
repo==0.3.0
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
ruamel.showoutput==0.1.7
s3transfer==0.7.0
safetensors==0.4.2
salesforce-lavis==1.0.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.13.0
segment-anything==1.0
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
snuggs==1.4.7
sounddevice==0.4.6
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spandrel==0.3.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.26.1
streamlit==1.28.1
supervision==0.16.0
svglib==1.5.1
sympy==1.12
tabulate==0.9.0
tb-nightly==2.11.0a20220906
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
text-unidecode==1.3
thinc==8.2.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.3
tomesd==0.1.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchdiffeq==0.2.3
torchmetrics==1.2.0
torchsde==0.2.5
torchvision==0.16.0+cu121
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trampoline==0.1.2
transformers==4.27.4
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.208
urllib3==2.0.7
uvicorn==0.24.0.post1
validators==0.22.0
virtualenv==20.24.6
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.9
weasel==0.3.4
webdataset==0.2.73
webencodings==0.5.1
websockets==11.0.3
webuiapi==0.9.8
Werkzeug==2.2.2
whichcraft==0.6.1
win32-setctime==1.1.0
wsproto==1.2.0
xformers==0.0.22.post7
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
